Move files to headed project where possible
