+++ game-app/game-core/src/main/java/games/strategy/engine/framework/GameRunner.java
@@ -1,39 +1,13 @@
 package games.strategy.engine.framework;
 
-import static games.strategy.engine.framework.CliProperties.LOBBY_GAME_COMMENTS;
-import static games.strategy.engine.framework.CliProperties.LOBBY_URI;
-import static games.strategy.engine.framework.CliProperties.SERVER_PASSWORD;
-import static games.strategy.engine.framework.CliProperties.TRIPLEA_CLIENT;
-import static games.strategy.engine.framework.CliProperties.TRIPLEA_GAME;
-import static games.strategy.engine.framework.CliProperties.TRIPLEA_HOST;
-import static games.strategy.engine.framework.CliProperties.TRIPLEA_MAP_DOWNLOAD;
-import static games.strategy.engine.framework.CliProperties.TRIPLEA_NAME;
-import static games.strategy.engine.framework.CliProperties.TRIPLEA_PORT;
-import static games.strategy.engine.framework.CliProperties.TRIPLEA_SERVER;
+import static games.strategy.engine.framework.CliProperties.*;
 
-import games.strategy.engine.auto.update.UpdateChecks;
-import games.strategy.engine.framework.map.download.DownloadMapsWindow;
-import games.strategy.engine.framework.startup.mc.ServerModel;
-import games.strategy.engine.framework.startup.ui.panels.main.HeadedServerSetupModel;
-import games.strategy.engine.framework.startup.ui.panels.main.game.selector.GameSelectorModel;
-import games.strategy.engine.framework.ui.MainFrame;
-import games.strategy.ui.Util;
-import java.awt.Component;
-import java.awt.Frame;
 import java.net.URI;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import javax.swing.SwingUtilities;
 import org.triplea.domain.data.UserName;
 import org.triplea.game.ApplicationContext;
-import org.triplea.java.Interruptibles;
-import org.triplea.java.ThreadRunner;
 import org.triplea.lobby.common.GameDescription;
-import org.triplea.swing.SwingAction;
-import org.triplea.util.ExitStatus;
 import org.triplea.util.Services;
 
 /**
@@ -44,72 +18,6 @@ public final class GameRunner {
   public static final String TRIPLEA_HEADLESS = "triplea.headless";
   public static final int PORT = 3300;
   public static final int MINIMUM_CLIENT_GAMEDATA_LOAD_GRACE_TIME = 20;
-
-  private static final GameSelectorModel gameSelectorModel = new GameSelectorModel();
-  private static HeadedServerSetupModel headedServerSetupModel;
-
-  private GameRunner() {}
-
-  /**
-   * Starts a new UI-enabled game client. This method will return before the game client UI exits.
-   * The game client UI will continue to run until it is shut down by the user.
-   *
-   * <p>No command-line arguments will launch a client; additional arguments can be supplied to
-   * specify additional behavior.
-   *
-   * @throws IllegalStateException If called from a headless environment.
-   */
-  public static void start() {
-    SwingUtilities.invokeLater(
-        () -> {
-          headedServerSetupModel = new HeadedServerSetupModel(gameSelectorModel);
-          MainFrame.buildMainFrame(headedServerSetupModel, gameSelectorModel);
-          headedServerSetupModel.showSelectType();
-          ThreadRunner.runInNewThread(GameRunner::showMainFrame);
-        });
-
-    UpdateChecks.launch();
-  }
-
-  /**
-   * Sets the 'main frame' to visible. In this context the main frame is the initial welcome (launch
-   * lobby/single player game etc..) screen presented to GUI enabled clients.
-   */
-  public static void showMainFrame() {
-    GameShutdownRegistry.runShutdownActions();
-
-    if (System.getProperty(TRIPLEA_SERVER, "false").equals("true")) {
-      MainFrame.show();
-      gameSelectorModel.loadDefaultGameSameThread();
-      final ServerModel serverModel = headedServerSetupModel.showServer();
-      MainFrame.addQuitAction(serverModel::cancel);
-      System.clearProperty(TRIPLEA_SERVER);
-    } else if (System.getProperty(TRIPLEA_CLIENT, "false").equals("true")) {
-      MainFrame.show();
-      headedServerSetupModel.showClient();
-      System.clearProperty(TRIPLEA_CLIENT);
-    } else {
-      final String saveGameFileName = System.getProperty(TRIPLEA_GAME, "");
-      if (!saveGameFileName.isEmpty()) {
-        final Path saveGameFile = Path.of(saveGameFileName);
-        if (Files.exists(saveGameFile) && !gameSelectorModel.load(saveGameFile)) {
-          // abort launch if we failed to load the specified game
-          return;
-        }
-      }
-      MainFrame.show();
-      gameSelectorModel.loadDefaultGameSameThread();
-      openMapDownloadWindowIfDownloadScheduled();
-    }
-  }
-
-  private static void openMapDownloadWindowIfDownloadScheduled() {
-    final String downloadableMap = System.getProperty(TRIPLEA_MAP_DOWNLOAD, "");
-    if (!downloadableMap.isEmpty()) {
-      SwingUtilities.invokeLater(
-          () -> DownloadMapsWindow.showDownloadMapsWindowAndDownload(downloadableMap));
-    }
-  }
 
   /** Spawns a new process to host a network game. */
   public static void hostGame(
@@ -153,25 +61,5 @@ public final class GameRunner {
     commands.add(prefix + TRIPLEA_NAME + "=" + userName.getValue());
     commands.add(Services.loadAny(ApplicationContext.class).getMainClass().getName());
     ProcessRunnerUtil.exec(commands);
-  }
-
-  public static void exitGameIfNoWindowsVisible() {
-    // Invoke later to add this check to the end of the event dispatcher queue
-    // and allow any potential in-flight 'setVisible' invocations to execute first.
-    SwingUtilities.invokeLater(
-        () -> {
-          final boolean allFramesClosed =
-              Arrays.stream(Frame.getFrames()).noneMatch(Component::isVisible);
-          if (allFramesClosed) {
-            ExitStatus.SUCCESS.exit();
-          }
-        });
-  }
-
-  /** After the game has been left, call this. */
-  public static void clientLeftGame() {
-    Util.ensureNotOnEventDispatchThread();
-    Interruptibles.await(() -> SwingAction.invokeAndWait(headedServerSetupModel::showSelectType));
-    showMainFrame();
   }
 }
