+++ game-app/game-headed/build.gradle
@@ -15,8 +15,10 @@ ext {
 dependencies {
     implementation "org.pushing-pixels:radiance-substance:$substanceVersion"
     implementation project(":game-app:ai")
+    implementation project(":game-app:domain-data")
     implementation project(":game-app:game-core")
     implementation project(":game-app:map-data")
+    implementation project(":http-clients:lobby-client")
     implementation project(":lib:java-extras")
     implementation project(":lib:swing-lib")
     testImplementation "org.sonatype.goodies:goodies-prefs:$sonatypeGoodiesPrefsVersion"
