+++ game-app/game-headed/src/main/java/games/strategy/engine/framework/startup/mc/HeadedLaunchAction.java
@@ -33,6 +33,7 @@ import javax.swing.SwingUtilities;
 import lombok.extern.slf4j.Slf4j;
 import org.triplea.domain.data.UserName;
 import org.triplea.game.chat.ChatModel;
+import org.triplea.game.client.HeadedGameRunner;
 import org.triplea.game.server.HeadlessGameServer;
 import org.triplea.java.Interruptibles;
 import org.triplea.sound.ClipPlayer;
@@ -78,7 +79,8 @@ public class HeadedLaunchAction implements LaunchAction {
       final IGame game,
       final Set<Player> players,
       final Chat chat) {
-    final TripleAFrame frame = TripleAFrame.create(game, localPlayers, chat);
+    final TripleAFrame frame =
+        TripleAFrame.create(game, localPlayers, chat, HeadedGameRunner::clientLeftGame);
 
     SwingUtilities.invokeLater(
         () -> {
